Rename configuration types file
